+++ apps/angular-kits-doc/stories/ui-components/form-field/chip-list/index.stories.ts
@@ -14,8 +14,8 @@ import { StoriesMeta, withThemeConfigProvider, withWrapperDecorator } from '@coc
 
 import descriptionMd from './description.md';
 
-export { Default } from './default.stories';
-export { Size } from './size.stories';
+export { Default } from './overview/default.stories';
+export { Size } from './overview/size.stories';
 
 const meta: StoriesMeta = {
   component: ChipListComponent,