+++ apps/angular-kits-doc/stories/ui-components/form-field/chip/overview/default.stories.ts
@@ -1,0 +1,58 @@
+import { ChipComponent } from '@cocokits/angular-form-field';
+import { AddonParametersControlType, CCK_CONTROL, renderWithPageTab } from '@cocokits/storybook-addon-theme';
+import { ngThemeArgsToTemplate, StoryObj } from '@cocokits/storybook-addon-theme-angular';
+
+export const Default: StoryObj<ChipComponent> = {
+  name: 'Default',
+  parameters: {
+    docs: {
+      description: {
+        story:
+          'Shows the default example with no additional configurations, providing an interactive example in its most basic form.',
+      },
+    },
+    cckAddon: {
+      renderConditions: [renderWithPageTab('Overview')],
+      source: [
+        {
+          filename: 'example.component.html',
+          language: 'angular-html',
+          code: `
+            <cck-chip
+              <% if (typeof type !== 'undefined') { %> type='<%= type %>' <% } %>
+              <% if (typeof size !== 'undefined') { %> size='<%= size %>' <% } %>
+              <% if (typeof color !== 'undefined') { %> color='<%= color %>' <% } %>
+              <% if (disabled) { %> disabled <% } %>
+              <% if (removable) { %> [removable]="true" <% } %>
+              >
+              <%= text %>
+            </cck-chip>
+          `,
+        },
+      ],
+      hasControl: true,
+      controls: [
+        CCK_CONTROL.text('Label'),
+        CCK_CONTROL.type(),
+        CCK_CONTROL.color(),
+        CCK_CONTROL.size(),
+        CCK_CONTROL.additional(),
+        CCK_CONTROL.removable(false),
+        CCK_CONTROL.disabled(false),
+      ],
+    },
+  },
+  render: (args) => ({
+    props: {
+      ...args,
+    },
+    template: `
+      <cck-chip
+        [removable]="cckControl.removable"
+        [disabled]="cckControl.disabled"
+        ${ngThemeArgsToTemplate(args)}>
+        {{cckControl.text}}
+      </cck-chip>
+    `,
+  }),
+};