+++ apps/angular-kits-doc/stories/ui-components/form-field/form-field/overview/chip-list.stories.ts
@@ -1,0 +1,139 @@
+import { FormFieldComponent } from '@cocokits/angular-form-field';
+import { AddonParametersControlType, CCK_CONTROL, renderWithPageTab } from '@cocokits/storybook-addon-theme';
+import { ngThemeArgsToTemplate, StoryObj } from '@cocokits/storybook-addon-theme-angular';
+
+export const ChipList: StoryObj<FormFieldComponent> = {
+  name: 'ChipList',
+  decorators: [
+    // withWrapperDecorator({}, {width: '300px'}),
+  ],
+  parameters: {
+    docs: {
+      description: {
+        story:
+          'Shows the default example with no additional configurations, providing an interactive example in its most basic form.',
+      },
+    },
+    cckAddon: {
+      renderConditions: [renderWithPageTab('Overview')],
+      source: [
+        {
+          filename: 'example.component.html',
+          language: 'angular-html',
+          code: `
+            <cck-form-field
+              <% if (typeof type !== 'undefined') { %> type='<%= type %>' <% } %>
+              <% if (typeof size !== 'undefined') { %> size='<%= size %>' <% } %>
+              <% if (typeof color !== 'undefined') { %> color='<%= color %>' <% } %>
+              <% if (disabled) { %> disabled <% } %>
+            >
+              <% if (label) { %>
+                <cck-label><%= label %></cck-label>
+              <% } %>
+              <% if (leading) { %>
+                <cck-leading><%= leading %></cck-leading>
+              <% } %>
+              <% if (prefixIcon !== 'none') { %>
+                <cck-prefix>
+                  <cck-svg-icon [icon]="YOUR_ICON"></cck-svg-icon>
+                </cck-prefix>
+              <% } %>
+
+              <cck-chip-list
+                [chips]="['Steak', 'Pizza', 'Burger']"
+                <% if (placeholder) { %> placeholder="<%= placeholder %>" <% } %>
+                <% if (addOnBlur) { %> [addOnBlur]="<%= addOnBlur %>" <% } %>
+              />
+
+              <% if (suffixIcon !== 'none') { %>
+                <cck-suffix>
+                  <cck-svg-icon [icon]="YOUR_ICON"></cck-svg-icon>
+                </cck-suffix>
+              <% } %>
+              <% if (trailing) { %>
+                <cck-trailing><%= trailing %></cck-trailing>
+              <% } %>
+              <% if (hint) { %>
+                <cck-hint><%= hint %></cck-hint>
+              <% } %>
+              <% if (error) { %>
+                <cck-hint [force]="true"><%= error %></cck-hint>
+              <% } %>
+            </cck-form-field>
+          `,
+        },
+      ],
+      hasControl: true,
+      controls: [
+        CCK_CONTROL.label('Label'),
+        CCK_CONTROL.placeholder('Placeholder'),
+        CCK_CONTROL.hint(''),
+        CCK_CONTROL.error(''),
+        CCK_CONTROL.leading(''),
+        CCK_CONTROL.trailing(''),
+        CCK_CONTROL.type(),
+        CCK_CONTROL.color(),
+        CCK_CONTROL.size(),
+        CCK_CONTROL.additional(),
+        CCK_CONTROL.prefixIcon('none'),
+        CCK_CONTROL.suffixIcon('none'),
+        CCK_CONTROL.disabled(),
+        {
+          displayName: 'Add On Blur',
+          default: false,
+          storyArgKey: 'addOnBlur',
+          type: AddonParametersControlType.Boolean,
+        },
+      ],
+    },
+  },
+  render: (args) => ({
+    props: {
+      ...args,
+      chips: ['Steak', 'Pizza', 'Burger'],
+    },
+    template: `
+      <cck-form-field
+        style="width: 100%;"
+        [disabled]="cckControl.disabled"
+        ${ngThemeArgsToTemplate(args)}>
+
+        @if(cckControl.label) {
+          <cck-label>{{cckControl.label}}</cck-label>
+        }
+
+        @if(cckControl.leading) {
+          <cck-leading>{{cckControl.leading}}</cck-leading>
+        }
+
+        @if(cckControl.prefixIcon !== 'none') {
+          <cck-prefix>
+            <cck-svg-icon [icon]="cckIcons[cckControl.prefixIcon]"></cck-svg-icon>
+          </cck-prefix>
+        }
+
+        <cck-chip-list
+          [chips]="chips"
+          placeholder="{{cckControl.placeholder}}"
+          [addOnBlur]="cckControl.addOnBlur"/>
+
+        @if(cckControl.suffixIcon !== 'none') {
+          <cck-suffix>
+            <cck-svg-icon [icon]="cckIcons[cckControl.suffixIcon]"></cck-svg-icon>
+          </cck-suffix>
+        }
+
+        @if(cckControl.trailing) {
+          <cck-trailing>{{cckControl.trailing}}</cck-trailing>
+        }
+
+        @if(cckControl.hint) {
+          <cck-hint>{{cckControl.hint}}</cck-hint>
+        }
+        @if(cckControl.error) {
+          <cck-error [force]="true">{{cckControl.error}}</cck-error>
+        }
+      </cck-form-field>
+    `,
+  }),
+};