+++ apps/angular-kits-doc/stories/ui-components/form-field/input/index.stories.ts
@@ -5,10 +5,7 @@ import { StoriesMeta, withThemeConfigProvider } from '@cocokits/storybook-addon-
 
 import descriptionMd from './description.md';
 
-export { Default } from './default.stories';
-// export { NativeType } from './native-type.stories';
-// export { NgModel } from './ng-model.stories';
-// export { ReactiveForm } from './reactive-form.stories';
+export { Default } from './overview/default.stories';
 
 const meta: StoriesMeta = {
   component: InputComponent,