+++ apps/angular-kits-doc/stories/ui-components/toggle/toggle/index.stories.ts
@@ -6,13 +6,10 @@ import { StoriesMeta, withThemeConfigProvider } from '@cocokits/storybook-addon-
 
 import descriptionMd from './description.md';
 
-export { Default } from './default.stories';
-export { Type } from './type.stories';
-export { Size } from './size.stories';
-export { Color } from './color.stories';
-// export { ThemeCocokitsBox } from './theme-cocokits-box.stories';
-// export { NgModel } from './ng-model.stories';
-// export { ReactiveForm } from './reactive-form.stories';
+export { Default } from './overview/default.stories';
+export { Type } from './overview/type.stories';
+export { Size } from './overview/size.stories';
+export { Color } from './overview/color.stories';
 
 const meta: StoriesMeta = {
   component: ToggleComponent,