+++ packages/react/button/vite.config.ts
@@ -2,14 +2,17 @@
 import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
 import react from '@vitejs/plugin-react-swc';
 import * as path from 'path';
+import copy from 'rollup-plugin-copy';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
+
 import { getExternalPackages } from '../../../tools/scripts/get-external-packages';
 
 export default defineConfig({
   root: __dirname,
   cacheDir: '../../../node_modules/.vite/packages/react/button',
+  assetsInclude: path.join(__dirname, 'README.md'),
 
   plugins: [
     react(),
@@ -19,6 +22,10 @@ export default defineConfig({
       entryRoot: 'src',
       tsconfigPath: path.join(__dirname, 'tsconfig.lib.json'),
       aliasesExclude: [/^@cocokits\//],
+    }),
+    copy({
+      targets: [{ src: 'README.md', dest: '../../../dist/packages/react/button' }],
+      hook: 'writeBundle',
     }),
   ],