+++ packages/react/tabs/vite.config.ts
@@ -2,6 +2,7 @@
 import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
 import react from '@vitejs/plugin-react-swc';
 import * as path from 'path';
+import copy from 'rollup-plugin-copy';
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 
@@ -18,6 +19,10 @@ export default defineConfig({
       entryRoot: 'src',
       tsconfigPath: path.join(__dirname, 'tsconfig.lib.json'),
       aliasesExclude: [/^@cocokits\//],
+    }),
+    copy({
+      targets: [{ src: 'README.md', dest: '../../../dist/packages/react/tabs' }],
+      hook: 'writeBundle',
     }),
   ],