+++ packages/angular/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cocokits/angular-utils
 
+## 1.0.3
+
+### Patch Changes
+
+- c9f1995: First Release 🔥
+- Updated dependencies [c9f1995]
+  - @cocokits/common-utils@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes