+++ packages/common/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cocokits/core
 
+## 1.1.1
+
+### Patch Changes
+
+- c9f1995: First Release 🔥
+- Updated dependencies [c9f1995]
+  - @cocokits/common-utils@1.0.3
+
 ## 1.1.0
 
 ### Minor Changes