+++ packages/internal/storybook-addon-theme-angular/CHANGELOG.md
@@ -1,1 +1,10 @@
 # @cocokits/storybook-addon-theme-angular
+
+## 0.0.1
+
+### Patch Changes
+
+- c9f1995: First Release 🔥
+- Updated dependencies [c9f1995]
+  - @cocokits/angular-core@1.1.1
+  - @cocokits/common-utils@1.0.3